+++ src/routes/login/first/+page.svelte
@@ -36,7 +36,7 @@
 	class="mx-auto mt-20 grid w-full max-w-sm space-y-6 rounded-lg bg-muted/5 p-3 py-6"
 	{...submit.attrs}
 >
-	<Logo class="mx-auto mt-auto w-52" aria-hidden />
+	<Logo class="mx-auto mt-auto" aria-hidden />
 
 	<p class="text-center">
 		{m.login_admin_introduction()}